refactor(quality): extract shared makePhaseRegistry factory across 12 subsystems
Collapse the duplicated registerPhase/getPhase/hasPhase trio -- identical in
every workflow subsystem except the lookup error label -- into a single
lib/make-phase-registry.js factory. Each dispatcher now calls
makePhaseRegistry('<label>') and wires only its own phases.

Removes 9 phase-registry.js entries from .quality-exceptions (the
duplicate-blocks were clones across this 12-file matrix). Behavior-preserving:
exports and the "No <label> phase handler registered" lookup error are
unchanged; the terse ci/pr-step variant adopts the standard
"must expose validate() and instructions()" message.

All 76 dispatcher/registry tests pass; full static-quality gate is clean.
