+++ .quality-exceptions
@@ -215,22 +215,13 @@ plugins/work/scripts/workflows/work-task-review/lib/phases/memorize.js
 plugins/work/scripts/listen-communication.js
 plugins/work/scripts/workflows/work-qa-feature-tester/lib/phases/memorize.js
 plugins/work/scripts/workflows/work-qa-feature-tester/lib/kind-checks/wiring.js
-plugins/work/scripts/workflows/work-spec/lib/phase-registry.js
 plugins/work/scripts/workflows/work-task-review/lib/phases/inputs.js
 plugins/work/scripts/workflows/work-qa-feature-tester/lib/phases/inputs.js
 plugins/work/scripts/workflows/work/steps/pr.js
 plugins/work/scripts/workflows/work-completion-checker/lib/kind-checks/fullstack.js
-plugins/work/scripts/workflows/work-completion-checker/lib/phase-registry.js
 plugins/work/scripts/workflows/work-pr-reviewer/lib/phases/inputs.js
 plugins/work/scripts/workflows/work-qa-feature-tester/lib/kind-checks/fullstack.js
 plugins/work/scripts/workflows/work/steps/check.js
-plugins/work/scripts/workflows/work-qa-feature-tester/lib/phase-registry.js
-plugins/work/scripts/workflows/work-tasks/lib/phase-registry.js
-plugins/work/scripts/workflows/work-code-checker/lib/phase-registry.js
-plugins/work/scripts/workflows/work-pr-reviewer/lib/phase-registry.js
-plugins/work/scripts/workflows/work-task-review/lib/phase-registry.js
-plugins/work/scripts/workflows/work-reports/lib/phase-registry.js
-plugins/work/scripts/workflows/work-pr-step/lib/phase-registry.js
 plugins/work/scripts/workflows/work-completion-checker/lib/kind-checks/kind-registry.js
 plugins/work/scripts/workflows/work-spec/lib/kind-checks/kind-registry.js
 plugins/work/scripts/workflows/lib/pr-state.js
